Split EMT code into its own file
Also, fix linter warnings by commenting several exported symbols.
