+++ edge_multi_test.go
